+++ classes/Configuration/DefaultConfiguration.php
@@ -12,18 +12,18 @@ class DefaultConfiguration {
    */
   public static function initialize($application) {
     $projectRoot = self::getProjectRoot();
-    $application_directory_path = $projectRoot . $application;
-    $application_directory = realpath($application_directory_path);
-    if (!is_dir($application_directory)) {
-      throw new ApplicationPartMissingException('Application folder ' . htmlspecialchars($application_directory_path) . ' not found', 1410538265);
+    $applicationDirectoryPath = $projectRoot . $application;
+    $applicationDirectory = realpath($applicationDirectoryPath);
+    if (!is_dir($applicationDirectory)) {
+      throw new ApplicationPartMissingException('Application folder ' . htmlspecialchars($applicationDirectoryPath) . ' not found', 1410538265);
     }
-    $application_directory .= '/';
+    $applicationDirectory .= '/';
     Configuration::set('phpframework', 'project_root', $projectRoot);
-    Configuration::set('phpframework', 'db.migrator.directory', $application_directory . '_sql/');
+    Configuration::set('phpframework', 'db.migrator.directory', $applicationDirectory . '_sql/');
     Configuration::set('application', 'application', $application);
-    Configuration::set('application', 'application_directory', $application_directory);
-    Configuration::set('application', 'routes_file', $application_directory . 'routes.php');
-    Configuration::set('application', 'templates_directory', $application_directory . 'templates/');
+    Configuration::set('application', 'application_directory', $applicationDirectory);
+    Configuration::set('application', 'routes_file', $applicationDirectory . 'routes.php');
+    Configuration::set('application', 'templates_directory', $applicationDirectory . 'templates/');
     Configuration::set('phpframework', 'version', '1.4-dev');
   }