+++ classes/Mvc/Request.php
@@ -18,8 +18,8 @@ class Request {
    */
   protected $requestMethod;
 
-  public function __construct($request_method) {
-    $this->requestMethod = $request_method;
+  public function __construct($requestMethod) {
+    $this->requestMethod = $requestMethod;
     // Early exit when not defined where to read request values
     if(!array_key_exists($this->requestMethod, $this->valueSources)) {
       throw new MethodNotSupportedException('Getting request parameters of ' . $this->requestMethod . ' is not supported.', 1415273543);
@@ -29,25 +29,25 @@ class Request {
   /**
    * Returns a parameter value from the request sent by the client
    *
-   * @param string $parameter_name Name of the request parameter to deliver
-   * @param mixed $default_value Default value to deliver when parameter is not found
-   * @param boolean $use_strict_mode Whether to allow searching in other request types than the used one
+   * @param string $parameterName Name of the request parameter to deliver
+   * @param mixed $defaultValue Default value to deliver when parameter is not found
+   * @param boolean $useStrictMode Whether to allow searching in other request types than the used one
    * @return mixed
    * @throws MethodNotSupportedException when the parameter is read for a request type which is not supported by the framework
    */
-  public function get($parameter_name, $default_value = NULL, $use_strict_mode = FALSE) {
+  public function get($parameterName, $defaultValue = NULL, $useStrictMode = FALSE) {
 
     try {
-      return $this->readRequestParameter($parameter_name, $this->valueSources[$this->requestMethod]['default']);
+      return $this->readRequestParameter($parameterName, $this->valueSources[$this->requestMethod]['default']);
     } catch(ParameterNotFoundException $ex) {}
 
-    if(!$use_strict_mode && isset($this->valueSources[$this->requestMethod]['fallback'])) {
+    if(!$useStrictMode && isset($this->valueSources[$this->requestMethod]['fallback'])) {
       try {
-        return $this->readRequestParameter($parameter_name, $this->valueSources[$this->requestMethod]['fallback']);
+        return $this->readRequestParameter($parameterName, $this->valueSources[$this->requestMethod]['fallback']);
       } catch(ParameterNotFoundException $ex) {}
     }
 
-    return $default_value;
+    return $defaultValue;
   }
 
   /**
@@ -77,29 +77,29 @@ class Request {
   /**
    * Returns the value of an parameter for a specified source if the source is supported and the parameter exists
    *
-   * @param string $parameter_name Name of the request parameter to deliver
-   * @param string $value_source Name of the source to use when reading the parameter
+   * @param string $parameterName Name of the request parameter to deliver
+   * @param string $valueSource Name of the source to use when reading the parameter
    * @return mixed
    * @throws ValueSourceNotSupportedException when the name of the source is unsupported
    * @throws ParameterNotFoundException when the parameter is not found in the source to avoid collision with other value types
    */
-  protected function readRequestParameter($parameter_name, $value_source) {
-    $source = null;
-    if($value_source == 'get') {
+  protected function readRequestParameter($parameterName, $valueSource) {
+
+    if($valueSource == 'get') {
       $source = &$_GET;
-    } elseif($value_source == 'post') {
+    } elseif($valueSource == 'post') {
       $source = &$_POST;
     }
 
-    if($source === null) {
-      throw new ValueSourceNotSupportedException('The source ' . $value_source . ' is not supported to read values from.', 1415273682);
+    if(!isset($source)) {
+      throw new ValueSourceNotSupportedException('The source ' . $valueSource . ' is not supported to read values from.', 1415273682);
     }
 
-    if(array_key_exists($parameter_name, $source)) {
-      return $source[$parameter_name];
+    if(array_key_exists($parameterName, $source)) {
+      return $source[$parameterName];
     }
 
-    throw new ParameterNotFoundException('Parameter ' . $parameter_name . ' was not found in source ' . $value_source, 1415273689);
+    throw new ParameterNotFoundException('Parameter ' . $parameterName . ' was not found in source ' . $valueSource, 1415273689);
   }
 
 }